requirements.txt
altair
asttokens
attrs
beautifulsoup4
blinker
cachetools
certifi
charset-normalizer
click
colorama
comm
contourpy
cycler
debugpy
decorator
executing
filelock
fonttools
fsspec
gdown
gitdb
GitPython
idna
ipykernel
ipython
jedi
Jinja2
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mpmath
narwhals
nest-asyncio
networkx
numpy
opencv-python
packaging
pandas
parso
pillow
pip
platformdirs
prompt_toolkit
protobuf
psutil
pure_eval
py-cpuinfo
pyarrow
pydeck
Pygments
pyparsing
PySocks
python-dateutil
pytz
PyYAML
pyzmq
referencing
requests
rich
rpds-py
scipy
seaborn
setuptools
six
smmap
soupsieve
stack-data
streamlit
sympy
tenacity
toml
torch
torchvision
tornado
tqdm
traitlets
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
utils
watchdog
wcwidth
joblib
tensorflow
scikit-learn
ultralytics
#playsound
#PyObjC
pydub 
ffmpeg
tf-keras